Arreglo de errores en buscar x id usuario

+++ src/main/java/org/nisum/nisumapi/model/User.java
@@ -25,7 +25,7 @@ public class User implements Serializable {
 
 	@Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Integer idUser;
+    private Long idUser;
 
     @NonNull
     private String name;
